vegeDetailに型定義を実装
